database restructures
