+++ invisible_backend/Cargo.toml
@@ -8,9 +8,9 @@ edition = "2021"
 name = "server"
 path = "src/server/server.rs"
 
-# [[bin]]
-# name = "client"
-# path = "src/server/client.rs"
+[[bin]]
+name = "main"
+path = "src/main.rs"
 
 
 [lib]
@@ -50,7 +50,7 @@ sled = "0.34.7"
 async-recursion = "1.0.2"
 starknet = { path = "crates/starknet-rs" }
 reqwest = "0.11.17"
-jsonwebtoken = "8.3.0"
+jsonwebtoken = "9.1.0"
 
 
 [build-dependencies]
