+++ invisible_backend/src/order_tab/db_updates.rs
@@ -5,12 +5,12 @@ use parking_lot::Mutex;
 use firestore_db_and_auth::ServiceSession;
 
 use crate::utils::{
-    firestore::{
+    notes::Note,
+    storage::firestore::{
         start_add_note_thread, start_add_order_tab_thread, start_delete_note_thread,
         start_delete_order_tab_thread,
     },
-    notes::Note,
-    storage::BackupStorage,
+    storage::local_storage::BackupStorage,
 };
 
 use super::OrderTab;
