+++ invisible_backend/src/smart_contract_mms/helpers/db_updates.rs
@@ -9,12 +9,12 @@ use crate::{
     order_tab::OrderTab,
     perpetual::perp_position::PerpPosition,
     utils::{
-        firestore::{
+        notes::Note,
+        storage::firestore::{
             start_add_note_thread, start_add_order_tab_thread, start_add_position_thread,
             start_delete_note_thread, start_delete_order_tab_thread,
         },
-        notes::Note,
-        storage::BackupStorage,
+        storage::local_storage::BackupStorage,
     },
 };
