+++ invisible_backend/src/utils/mod.rs
@@ -1,7 +1,7 @@
 pub mod cairo_output;
 pub mod errors;
 pub mod ffi;
-pub mod firestore;
-pub mod notes;
+
 pub mod crypto_utils;
+pub mod notes;
 pub mod storage;
